Adding in horizontal limit to head rotation
